refactor: PostStatisticsDto 위치 이동
