+++ api/src/main/kotlin/org/deepforest/dcinside/post/dto/PostDto.kt
@@ -45,3 +45,20 @@ class PostRequestDto(
             title = title
         )
 }
+
+class PostStatisticsDto(
+    val viewCount: Long = 0L,
+    val likeCount: Long = 0L,
+    val dislikeCount: Long = 0L,
+    val commentCount: Long = 0L
+) {
+    companion object {
+        fun from(postStatistics: PostStatistics) =
+            PostStatisticsDto(
+                postStatistics.viewCount,
+                postStatistics.likeCount,
+                postStatistics.dislikeCount,
+                postStatistics.commentCount
+            )
+    }
+}
