--- api/src/main/kotlin/org/deepforest/dcinside/post/dto/PostStatisticsDto.kt
@@ -1,20 +1,0 @@
-package org.deepforest.dcinside.post.dto
-
-import org.deepforest.dcinside.entity.post.PostStatistics
-
-class PostStatisticsDto(
-    val viewCount: Long = 0L,
-    val likeCount: Long = 0L,
-    val dislikeCount: Long = 0L,
-    val commentCount: Long = 0L
-) {
-    companion object {
-        fun from(postStatistics: PostStatistics) =
-            PostStatisticsDto(
-                postStatistics.viewCount,
-                postStatistics.likeCount,
-                postStatistics.dislikeCount,
-                postStatistics.commentCount
-            )
-    }
-}
